Changes to what is displayed on the admin teams and players tab for Admins and Coordinator
For Admins, on admin teams view, only display teams for the currently selected league. On players view, only display players in the currently selected league.

For Coordinators, on admin teams view, only display teams for the league they coordinate for. On admin players view, only display players in the league they coordinate for