When creating player:
If user type is admin, allow player to be created for teams in currently selected league.
If user type is coordinator, allow player to be created for teams in league the user coordinates for.
If user type is manager, only allow player to be created for team the user manages